evathon/backend/requirements.txt
pyserial>=3.5
flask>=3.0.0
flask-cors>=4.0.0
